Software-Domain - the domain of first choice for software
Software forms the building blocks of the world. We see Software-Domains at the forefront in this next stage of the revolution. It will be the place online that unifies this disparate group of technologically connected engineers, developers, retailers and purchasers. That brings them together as a profession, as a group of creatives, as visionaries. Setting them apart within their individual professions and as members of one of the most integral sectors that exists today, and into the future.

The Software-Domain is literally speaking to this market in their own language. It is made up of tech savvy, forward-thinking companies and leaders eager to align themselves with a web presence that identifies them as industry professionals. By the nature of their work, they will be easy to reach. Software is developed as a communal effort by the creative minds behind computer programs. Some develop applications and apps for computers and other devices. Others create the underlying systems that run devices or control networks. More still roll out these programs and apps into the world through online and offline retail outlets.

Software is big business across the world – and it changes rapidly. As we move from databases to the cloud into the Internet of Things, as software links everything in our lives from healthcare to refrigerators, automobiles to mobile phones, there has never been a better time to be involved in the business of software.

Software development, as an occupation in the US alone, is projected to grow 22% from 2012 to 2022, faster than the average for all other occupations. The main reason for this rapid growth is a large increase in the demand for computer software. As more and more people enter this disparate, specialized occupation, finding a home online for like-minded individuals will become more and more important.

The immediate opportunity for the extremely specific Software-Domain is its appeal to the large number of sole proprietor, small software developers and consultants that comprise almost 15,000 companies in the US.

Developers of apps for iPhone, Android and other mobile platforms make up another large group who will look to a Software-Domain for marketing. It will also attract valuable premiums for generic terms and high Cost Per Click words for domain investors.

Marc Mueller
http://www.domainregistry.de/software-domain.html